Homeworks 1, 3 and 4
- Add images created for the first homework. One is created in
Cinema4D and the other in Blender, both from scratch, modeling
from primitive shapes, with added materials (from scratch as well)
and with added dome/sun with sky lights.
- Add code for vectors with vector operations and rays.
- Create an image for the third task with calculating rays from
the camera to the image plane and creating colors based off that.
- Add code for triangles.
- Add functions for printing vectors and triengles to an output
stream, so that it's easier to represent the fourth task.
- Separate the code into different files so that it's easier to
read and write.